% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGIFA-package.R
\docType{package}
\name{TGIFA-package}
\alias{TGIFA}
\alias{TGIFA-package}
\title{TGIFA: A Missing Data Imputation Method for Metabolomics Data Using a Truncated Gaussian Infinite Factor Model}
\description{
A missing data imputation method using a truncated infinite factor analysis model designed for high-dimensional metabolomics data.
}
\author{
\strong{Maintainer}: Kate Finucane \email{kate.finucane@ucdconnect.ie} [copyright holder]

Authors:
\itemize{
  \item Isobel Claire Gormley
  \item Lorraine Brennan
}

}
\keyword{internal}
