% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGIFA_model.R
\name{TGIFA_model}
\alias{TGIFA_model}
\title{Truncated Gaussian Infinite Factor Analysis Imputation Model}
\usage{
TGIFA_model(
  input_data,
  coding = NA,
  n.iters = 10000,
  k.star = 5,
  verbose = TRUE,
  burn = 5000,
  thin = 5,
  kappa_1 = 3L,
  kappa_2 = 2L,
  a_sigma = 1L,
  b_sigma = 0.25,
  a_1 = 2.1,
  a_2 = 3.1,
  prop_mult = list(mu_prop_mult = 0.05, eta_prop_mult = 3L, lambda_prop_mult = 3L,
    sigma_inv_prop_mult = 1.5),
  return_chain = FALSE,
  cred_level = 0.95,
  var_mult = 1
)
}
\arguments{
\item{input_data}{The dataset with missing value requiring imputation in matrix format.}

\item{coding}{The coding used to indicate a data entry is missing. Defaults to \code{NA}.}

\item{n.iters}{The number of MCMC iterations desired. Defaults to 10000.}

\item{k.star}{Parameter \eqn{k^{*}} in the TGIFA model. Defaults to 5.}

\item{verbose}{Is a readout desired? Defaults to \code{TRUE}.}

\item{burn}{The number of MCMC iterations to be discarded in an initial burn. Defaults to 5000.}

\item{thin}{The level of thinning to take place in the MCMC chain. Defaults to 5, meaning only every fifth draw will be kept.}

\item{kappa_1}{Parameter \eqn{\kappa_1} in the TGIFA model. Defaults to 3.}

\item{kappa_2}{Parameter \eqn{\kappa_2} in the TGIFA model. Defaults to 2.}

\item{a_sigma}{Parameter \eqn{a_{\sigma}} in the TGIFA model. Defaults to 1.}

\item{b_sigma}{Parameter \eqn{b_{\sigma}} in the TGIFA model. Defaults to 0.25.}

\item{a_1}{Parameter \eqn{a_1} in the TGIFA model. Defaults to 2.1.}

\item{a_2}{Parameter \eqn{a_2} in the TGIFA model. Defaults to 3.1.}

\item{prop_mult}{A list containing scalar multipliers for the covariance of parameters with Metropolis Hastings or Exchange algorithm update steps.}

\item{return_chain}{Should resulting draws from the MCMC chain be returned? Defaults to \code{FALSE}.}

\item{cred_level}{The desired credible interval to report for imputed values. E.g., \code{cred_level = 0.95} will return a 95 percent credible interval.}

\item{var_mult}{The desired proportion of the variance of the input data to be assigned to initial values of the idiosyncratic errors.}
}
\value{
Returns a list containing entries:
\itemize{
\item imputed_dataset: The imputed dataset in matrix format.
\item imputation_info: A dataframe containing information on imputed values and associated uncertainty for each missing entry in the original dataset.
\item chain: Returned only if return_chain = \code{TRUE}. Provides all draws from the MCMC chain and further information.
\itemize{
\item store_data: MCMC draws of the imputed data entries.
\item store_alpha: MCMC draws of the \eqn{\alpha} parameter.
\item store_sigma_inv: MCMC draws of the \eqn{\sigma^{-2}} entries.
\item store_lambda: MCMC draws of the loadings matrix.
\item store_eta: MCMC draws of the latent factor scores.
\item store_mu: MCMC draws of the mean vector.
\item store_k_t: MCMC draws of the effective factor number (unchanging).
\item store_phi: MCMC draws of the local shrinkage parameters.
\item store_delta: MCMC draws of the multiplicative gamma process parameters.
\item store_b_sigma: MCMC draws of the \eqn{b_{\sigma}} parameter (unchanging).
\item store_Z: MCMC draws of the missingness designation indicator for each imputed entry.
\item input_data: The dataset input into the tIFA model.
\item input_missingness: The missingness pattern input into the tIFA model.
\item store_MH_mu: Exchange algorithm acceptance for the mean parameter.
\item store_MH_lambda: Exchange algorithm acceptance for the loadings matrix..
\item store_MH_eta: Metropolis-Hastings acceptance for the latent factor scores
\item store_MH_sigma_inv: Exchange algorithm acceptance for the \eqn{\sigma^{-2}} entries.
}
}
}
\description{
Applies the TGIFA imputation method to the input dataset.
}
\examples{
# generate example data
set.seed(1)
example_data <- matrix(abs(rnorm(200)), nrow = 5)

# add missingness to example data coded as 0.001
example_data[4, 2] <- 0.001
example_data[2, 18] <- 0.001

# run TGIFA model
# short chain for example
res <- TGIFA_model(input_data = example_data, coding = 0.001, n.iters = 100, k.star = 3,
                  burn = 50, thin = 5)

# checkout imputed dataset
res$imputed_dataset

# checkout further information on imputed entries
res$imputation_info

}
